requirements.txt
asgiref==3.8.1
attrs==23.2.0
dj-rest-auth==6.0.0
Django==5.0.4
django-admin-interface==0.28.6
django-colorfield==0.11.0
django-extensions==3.2.3
django-allauth==0.62.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
pillow==10.3.0
PyJWT==2.8.0
python-slugify==8.0.4
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.0
sqlparse==0.5.0
text-unidecode==1.3
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.6.0